move some stuff to mock package
