+++ backend/module/chaos/serverexperimentation/xds/xds_cache_test.go
@@ -2,6 +2,7 @@ package xds
 
 import (
 	"context"
+	"github.com/lyft/clutch/backend/mock/service/chaos/experimentation/experimentstoremock"
 	"testing"
 	"time"
 
@@ -17,7 +18,6 @@ import (
 
 	experimentation "github.com/lyft/clutch/backend/api/chaos/experimentation/v1"
 	serverexperimentation "github.com/lyft/clutch/backend/api/chaos/serverexperimentation/v1"
-	xds_testing "github.com/lyft/clutch/backend/module/chaos/serverexperimentation/xds/testing"
 )
 
 const (
@@ -241,7 +241,7 @@ func TestSetSnapshotV3WithTTL(t *testing.T) {
 }
 
 func TestRefreshCache(t *testing.T) {
-	s := xds_testing.MockStorer{}
+	s := experimentstoremock.MockStorer{}
 	rtdsConfig := RTDSConfig{
 		layerName:     "test_layer",
 		ingressPrefix: "ingress",
